refactor: restructure files and folders
