+++ src/cli/bin.js
@@ -1,5 +1,5 @@
 #!/usr/bin/env node
 const args = process.argv.slice(2);
 const cli = require('./index');
-cli.main(args);
+cli.run(args);
 process.title = ["swa", ...args].join(" ");
