+++ src/core/utils/cli.spec.ts
@@ -1,4 +1,4 @@
-jest.mock("./constants", () => {});
+jest.mock("../constants", () => {});
 import mockFs from "mock-fs";
 import path from "path";
 import { argv, createStartupScriptCommand } from "./cli";
