+++ src/core/utils/logger.ts
@@ -1,7 +1,7 @@
 import chalk from "chalk";
 import type http from "http";
 import { DEFAULT_CONFIG } from "../../config";
-import { SWA_CLI_APP_PROTOCOL } from "./constants";
+import { SWA_CLI_APP_PROTOCOL } from "../constants";
 
 export const logger = {
   _print(prefix: string | null, data: string) {
