+++ src/msha/auth/routes/auth-login-provider.ts
@@ -1,11 +1,11 @@
 import { IncomingMessage } from "http";
-import { response } from "../../core";
+import { response } from "../../../core";
 
 const fs = require("fs").promises;
 const path = require("path");
 
 const httpTrigger = async function (context: Context, request: IncomingMessage) {
-  const body = await fs.readFile(path.join(__dirname, "..", "..", "public", "auth.html"), "utf-8");
+  const body = await fs.readFile(path.join(__dirname, "..", "..", "..", "public", "auth.html"), "utf-8");
   context.res = response({
     context,
     status: 200,
