+++ src/msha/handlers/error-page.handler.ts
@@ -1,7 +1,7 @@
 import type http from "http";
-import { logRequest } from "../core";
-import { SWA_CLI_APP_PROTOCOL } from "../core/utils/constants";
-import { responseOverrides } from "./routes-engine";
+import { logRequest } from "../../core";
+import { SWA_CLI_APP_PROTOCOL } from "../../core/constants";
+import { responseOverrides } from "../routes-engine";
 
 export function pageNotFoundResponse(
   req: http.IncomingMessage,
