+++ src/msha/routes-engine/rules/response-overrides.spec.ts
@@ -1,6 +1,6 @@
-jest.mock("../../../core/utils/constants", () => {});
+jest.mock("../../../core/constants", () => {});
 import { DEFAULT_CONFIG } from "../../../config";
-import { responseOverrides } from "./responseOverrides";
+import { responseOverrides } from "./response-overrides";
 
 describe("responseOverrides()", () => {
   let req: any;
